// apps/api/src/modules/contacts/dto/create-contact.dto.d.ts
export declare class CreateContactDto {
    name: string;
    email: string;
    phone?: string;
    companyName?: string;
}
export declare class UpdateContactDto {
    name?: string;
    email?: string;
    phone?: string;
    companyName?: string;
}
//# sourceMappingURL=create-contact.dto.d.ts.map